Move impl to lib/shared
