+++ lib/shared/package.json
@@ -31,10 +31,13 @@
     "diff": "^5.2.0",
     "dompurify": "^3.0.4",
     "highlight.js": "^10.7.3",
+    "ini": "^4.1.2",
     "isomorphic-fetch": "^3.0.0",
     "js-tiktoken": "^1.0.10",
     "lodash": "^4.17.21",
+    "lru-cache": "^10.0.0",
     "marked": "^4.0.16",
+    "re2-wasm": "^1.0.2",
     "vscode-uri": "^3.0.7"
   },
   "devDependencies": {
@@ -43,6 +46,7 @@
     "@types/diff": "^5.0.9",
     "@types/dompurify": "^3.0.2",
     "@types/highlight.js": "^9.12.4",
+    "@types/ini": "^4.1.0",
     "@types/isomorphic-fetch": "^0.0.39",
     "@types/lodash": "^4.14.195",
     "@types/marked": "^5.0.0",
