+++ lib/shared/src/cody-ignore/repo-name-resolver.ts
@@ -1,25 +1,30 @@
 import ini from 'ini'
 import * as vscode from 'vscode'
 
-import { convertGitCloneURLToCodebaseName, isFileURI } from '@sourcegraph/cody-shared'
-
-import { logDebug } from '../log'
-
 import { LRUCache } from 'lru-cache'
-import { gitRemoteUrlFromGitExtension } from './git-extension-api'
+import { isFileURI } from '../common/uri'
+import { logDebug } from '../logger'
+import { convertGitCloneURLToCodebaseName } from '../utils'
 
 export type RemoteUrlGetter = (uri: vscode.Uri) => Promise<string | undefined>
 type FsPath = string
 type RepoName = string
 
+type GitRemoteUrlFromGitExtension = (uri: vscode.Uri) => string | undefined
+
 export class RepoNameResolver {
     private platformSpecificGitRemoteGetters: RemoteUrlGetter[] = []
     private fsPathToRepoNameCache = new LRUCache<FsPath, RepoName>({ max: 1000 })
+    private gitRemoteUrlFromGitExtension: GitRemoteUrlFromGitExtension | undefined = undefined
 
     /**
      * Currently is used to set node specific remote url getters on the extension init.
      */
-    public init(platformSpecificGitRemoteGetters: RemoteUrlGetter[] = []) {
+    public init(
+        gitRemoteUrlFromGitExtension: GitRemoteUrlFromGitExtension,
+        platformSpecificGitRemoteGetters: RemoteUrlGetter[] = []
+    ) {
+        this.gitRemoteUrlFromGitExtension = gitRemoteUrlFromGitExtension
         this.platformSpecificGitRemoteGetters = platformSpecificGitRemoteGetters
     }
 
@@ -42,7 +47,7 @@ export class RepoNameResolver {
         }
 
         try {
-            let remoteOriginUrl = gitRemoteUrlFromGitExtension(uri)
+            let remoteOriginUrl = this.gitRemoteUrlFromGitExtension?.(uri)
 
             if (!remoteOriginUrl) {
                 remoteOriginUrl = await gitRemoteUrlFromTreeWalk(uri)
