+++ vscode/package.json
@@ -1269,7 +1269,6 @@
     "diff": "^5.2.0",
     "fast-xml-parser": "^4.3.2",
     "glob": "^7.2.3",
-    "ini": "^4.1.2",
     "isomorphic-fetch": "^3.0.0",
     "js-levenshtein": "^1.1.6",
     "lexical": "^0.13.1",
@@ -1304,7 +1303,6 @@
     "@types/express": "^4.17.17",
     "@types/fs-extra": "^11.0.4",
     "@types/glob": "^8.0.0",
-    "@types/ini": "^4.1.0",
     "@types/isomorphic-fetch": "^0.0.39",
     "@types/js-levenshtein": "^1.1.1",
     "@types/lodash": "^4.14.195",
