+++ vscode/src/main.ts
@@ -16,6 +16,7 @@ import {
     setLogger,
 } from '@sourcegraph/cody-shared'
 
+import { repoNameResolver } from '@sourcegraph/cody-shared/src/cody-ignore/repo-name-resolver'
 import { ContextProvider } from './chat/ContextProvider'
 import type { MessageProviderOptions } from './chat/MessageProvider'
 import { ChatManager, CodyChatPanelViewType } from './chat/chat-view/ChatManager'
@@ -49,8 +50,7 @@ import { getChatModelsFromConfiguration, syncModelProviders } from './models/uti
 import type { FixupTask } from './non-stop/FixupTask'
 import { CodyProExpirationNotifications } from './notifications/cody-pro-expiration'
 import { showSetupNotification } from './notifications/setup-notification'
-import { gitAPIinit } from './repository/git-extension-api'
-import { repoNameResolver } from './repository/repo-name-resolver'
+import { gitAPIinit, gitRemoteUrlFromGitExtension } from './repository/git-extension-api'
 import { SearchViewProvider } from './search/SearchViewProvider'
 import { AuthProvider } from './services/AuthProvider'
 import { CharactersLogger } from './services/CharactersLogger'
@@ -327,7 +327,7 @@ const register = async (
 
     const commandsManager = platform.createCommandsProvider?.()
     setCommandController(commandsManager)
-    repoNameResolver.init(platform.getRemoteUrlGetters?.())
+    repoNameResolver.init(gitRemoteUrlFromGitExtension, platform.getRemoteUrlGetters?.())
 
     // Execute Cody Commands and Cody Custom Commands
     const executeCommand = (
